Divided code into header and source file
