+++ image_preprocessing/CMakeLists.txt
@@ -1,5 +1,4 @@
 cmake_minimum_required(VERSION 3.5)
-
 project(image_preprocessing)
 
 # Find dependencies
@@ -10,17 +9,23 @@ find_package(cv_bridge REQUIRED)
 find_package(image_transport REQUIRED)
 find_package(OpenCV REQUIRED)
 
+# Specify the path to your header file's directory
+set(INCLUDE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/home/rachit/percep_ws/src/ros-perception-pipeline/image_preprocessing/include)
+
+# Include the directory containing your header file
+include_directories(${INCLUDE_DIR})
+
 # Add executable
-add_executable(test_node src/test_node.cpp)
+add_executable(image_process_node src/image_process_node.cpp)
 
 # Include directories for the executable
-target_include_directories(test_node
+target_include_directories(image_process_node
   PRIVATE
   ${sensor_msgs_INCLUDE_DIRS}
 )
 
 # Link the executable to the required libraries
-ament_target_dependencies(test_node
+ament_target_dependencies(image_process_node
   rclcpp
   sensor_msgs
   cv_bridge
@@ -30,12 +35,14 @@ ament_target_dependencies(test_node
 
 # Install the executable
 install(TARGETS 
-  test_node
+  image_process_node
   DESTINATION lib/${PROJECT_NAME}
 )
 
 install(DIRECTORY launch config
 DESTINATION share/${PROJECT_NAME})
+
+include_directories(include)
 
 # Install launch files, config files, and other directories if necessary
 # install(DIRECTORY launch
